fix: Compilation on Windows
Changes to be committed:
	modified:   ECS/Systems/ChatBoxSystem.hpp
	modified:   ECS/Systems/Server/NetworkSystem.hpp
	modified:   R-Type-documentations
	modified:   includes/Packets.hpp
	modified:   includes/ZLib.hpp
	modified:   srcs/Client/include/EndMenu.hpp
	modified:   srcs/Client/include/Game.hpp
	modified:   srcs/Client/include/MainMenu.hpp
	modified:   srcs/Client/include/Network.hpp
	modified:   srcs/Client/src/EndMenu.cpp
	modified:   srcs/Client/src/Screen.cpp
	modified:   srcs/Server/src/Main.cpp

+++ srcs/Server/src/Main.cpp
@@ -9,7 +9,7 @@ int main()
 {
     try
     {
-        std::srand(std::time(nullptr));
+        std::srand(static_cast<std::uint32_t>(std::time(nullptr)));
         rtype::loopSystem game;
         game.runNetwork();
         game.runGame();
